:art: move `publish_event_on_websocket` to own module and make publish call when events are retreived from redis via pubsub
